Accept Config file
Allow config file to be passed
